Review: The DAG allows to switch between append-only and delete-load functionality

+++ airflow/plugins/operators/load_dimension.py
@@ -11,19 +11,28 @@ class LoadDimensionOperator(BaseOperator):
                  conn_id="",
                  table="",
                  load_sql_stmt="",
+                 append_data=True,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.conn_id = conn_id
         self.table = table
         self.load_sql_stmt = load_sql_stmt
+        self.append_data = append_data
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.conn_id)
         self.log.info(f"Loading dimension table {self.table} in Redshift")
+        
+        if self.append_data == False:
+            redshift.run(f"""
+                BEGIN;
+                TRUNCATE TABLE {self.table};
+                COMMIT;
+            """)
+
         redshift.run(f"""
             BEGIN;
-            TRUNCATE TABLE {self.table};
             INSERT INTO {self.table} 
             {self.load_sql_stmt};
             COMMIT;
